Fix unit tests
The unit tests were broken by a change to openstacklib
Id09c3358c5843510e6a2a8c0e2d4aeb3607e098b which renamed some of the
resources in the openstacklib classes. The problem is that our tests in
this module should not have been testing what occurs within openstacklib
so this fix updates the tests to only check what we are doing within
this module.

Change-Id: Ia3ec5a877d6e59520821fe6cd62d6b2f335d8b27
Depends-On: I9d535ab38afea852559df2b3073bd4b74a2a3947