Panko api wsgi changed path for Ubuntu
When installing the panko-api and panko-common
packages for Ubuntu the path to the wsgi script is
no longer /usr/share/panko-common/app.wsgi but is
now /usr/lib/python2.7/dist-packages/panko/api/app.wsgi

This patch changes that script path. It depends on the
Gnocchi change in order to test this since Gnocchi has
to be fixed in order to run testing. [1]

[1] https://review.openstack.org/#/c/548032/

Depends-On: I070f88dd40ca4129cbd3abf5b4f5ecb511b2895e
Change-Id: I066bd87ce839d380942adcb73fcaa6d915f801c1