Remove deprecation message from panko::api
The panko::api class was marked as deprecated[1] a long ago without
either any clear announcement or release note.
However in fact puppet-panko has been maitainced and Panko is still
a part of OpenStack projects now.
This patch removes the deprecation message regarding the fact that
there are no actual plans to remove it.

[1] 272f764ce3a54e0d9ae165fb3698b793945ae4b0

Change-Id: I9a64b0b6f6f883d15ab81ea7a8b7e2a18597b119